cancel button changed color, and alert pop up implemented

+++ The Redline/screens/EditEventsScreen.js
@@ -1,5 +1,5 @@
 import React, {useState, useEffect} from 'react'
-import { Dimensions, Button, View, Text, StyleSheet, ScrollView, TouchableOpacity } from 'react-native'
+import { Dimensions, Button, View, Text, StyleSheet, ScrollView, TouchableOpacity, Alert } from 'react-native'
 import CustomInput from '../components/CustomInput';
 import CustomButton from '../components/CustomButton';
 import EventDescriptionInput from '../components/EventDescriptionInput';
@@ -38,6 +38,8 @@ const EditEventsScreen = ({ route, navigation }) => {
     const [isValidDate, setIsValidDate] = useState(true);
     const [isValidTime, setIsValidTime] = useState(true);
     const [isValidDescription, setIsValidDescription] = useState(true);
+
+    const [showBox, setShowBox] = useState(true);
 
     //User Data
     const user = firebase.auth().currentUser;
@@ -223,13 +225,32 @@ const EditEventsScreen = ({ route, navigation }) => {
     }
 
     const onDeleteEventPressed = () => {
-        firebase.firestore().collection('events').doc(dataId)
-            .delete()
-            .then(() => {
-            console.log('Event deleted!');
-        });
 
-        navigation.navigate("Profile")
+        return Alert.alert(
+            "Are your sure?",
+            "Are you sure you want to delete your event?",
+            [
+              // The "Yes" button
+              {
+                text: "Yes",
+                onPress: () => {
+                    setShowBox(false);
+                    firebase.firestore().collection('events').doc(dataId)
+                        .delete()
+                        .then(() => {
+                        console.log('Event deleted!');
+                    });
+
+                    navigation.navigate("Profile")
+                },
+              },
+              // The "No" button
+              // Does nothing but dismiss the dialog when tapped
+              {
+                text: "No",
+              },
+            ]
+          );
     }
 
     // UI Components
@@ -323,7 +344,7 @@ const EditEventsScreen = ({ route, navigation }) => {
                         <CustomButton onPress={onSubmitPressed} buttonName="Submit" type="PRIMARY"/>
                     </View>
                     <View style={{flexDirection:"row", marginBottom: 0, marginTop: 0 }}>
-                        <CustomButton onPress={onDeleteEventPressed} buttonName="Delete Event" type="PRIMARY"/>
+                        <CustomButton onPress={onDeleteEventPressed} buttonName="Delete Event" type="RED"/>
                     </View>
                     <TouchableOpacity onPress={onCancelPressed}>
                         <Text style = {GlobalStyles.blueText} iconName="account-outline"> Cancel </Text>
